+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 IMPROVEMENTS:
 
 * Change resource handling to use locking mechanism when resource parallel handling is not supported by vCD. [GH-#255] 
-* **New Resource:** edge gateway `vcd_edgegateway` creates and deletes edge gateways - ([#262](https://github.com/terraform-providers/terraform-provider-vcd/issues/262))
+
+FEATURES:
+
+* **New Resource:** Load Balancer Service Monitor `vcd_lb_service_monitor` - [GH-256]
+* **New Data Source:** Load Balancer Service Monitor `vcd_lb_service_monitor` - [GH-256]
+* **New Resource:** edge gateway `vcd_edgegateway` creates and deletes edge gateways - [GH-262](https://github.com/terraform-providers/terraform-provider-vcd/issues/262)
 
 ## 2.3.0 (May 29, 2019)